RUN-211(1): Plan for refactoring Library Templates(file splitting) (#171)
* Create LibraryCompose

* changed

* Splitting files into subsections

* Update  paddingValue parameter
